requirements.txt
flask
flask-socketio
paho-mqtt
eventlet